@@ -14,16 +14,18 @@
     "test": "lerna run test",
     "build": "lerna run build",
     "build:client": "lerna run build:client",
-    "watch": "lerna run --parallel watch"
+    "watch": "lerna run --parallel watch",
+    "sync:version": "syncpack fix-mismatches"
   },
   "author": "",
   "devDependencies": {
     "@types/express": "^4.17.15",
     "@types/express-ws": "^3.0.1",
+    "syncpack": "^8.4.11",
     "ts-jest": "^29.0.3",
-    "typescript": "^4.3.5"
+    "typescript": "^4.9.4"
   },
   "dependencies": {
-    "lerna": "^6.3.0"
+    "lerna": "^6.4.0"
   }
 }
