@@ -8,8 +8,7 @@
   },
   "devDependencies": {
     "@types/jest": "^26.0.24",
-    "@types/node": "^16.4.9",
-    "typescript": "^4.3.5"
-  },
-  "dependencies": {}
+    "@types/node": "^18.11.18",
+    "typescript": "^4.9.4"
+  }
 }
