@@ -13,19 +13,19 @@
   "devDependencies": {
     "@types/axios": "^0.14.0",
     "@types/jest": "^26.0.24",
-    "@types/node": "^16.4.9",
+    "@types/node": "^18.11.18",
     "@types/plotly.js": "^1.54.14",
     "@types/plotly.js-dist-min": "^2.3.0",
-    "@types/react": "^17.0.15",
+    "@types/react": "^18.0.26",
     "@types/react-color": "^3.0.6",
-    "@types/react-dom": "^17.0.9",
+    "@types/react-dom": "^18.0.10",
     "@types/react-router-dom": "^5.1.8",
     "@types/styled-components": "^5.1.11",
     "compression-webpack-plugin": "^9.0.0",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.2",
     "ts-loader": "^9.2.4",
-    "typescript": "^4.3.5",
+    "typescript": "^4.9.4",
     "webpack": "^5.48.0",
     "webpack-cli": "^4.7.2",
     "webpack-dev-server": "^3.11.2"
@@ -33,8 +33,9 @@
   "dependencies": {
     "@ftl/player": "1.0.0",
     "@ftl/protocol": "1.0.0",
+    "@beyond/protocol": "^0.1.5",
     "@ftl/stream": "1.0.0",
-    "axios": "^0.21.4",
+    "axios": "^1.2.2",
     "formik": "^2.2.9",
     "is-arguments": "^1.1.1",
     "is-generator-function": "^1.0.9",
@@ -42,14 +43,14 @@
     "plotly.js-dist-min": "^2.3.1",
     "pupa": "^3.0.0",
     "query-string": "^7.0.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-color": "^2.19.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.2.0",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "recoil": "^0.4.0",
-    "styled-components": "^5.3.0",
+    "react-router": "^6.6.1",
+    "react-router-dom": "^6.6.1",
+    "recoil": "^0.7.6",
+    "styled-components": "^5.3.6",
     "which-typed-array": "^1.1.5"
   }
 }
