@@ -8,18 +8,18 @@
   },
   "devDependencies": {
     "@types/jest": "^26.0.24",
-    "@types/node": "^16.4.9",
-    "typescript": "^4.3.5"
+    "@types/node": "^18.11.18",
+    "typescript": "^4.9.4"
   },
   "dependencies": {
     "@ftl/types": "1.0.0",
     "@ftl/api": "1.0.0",
-    "@tsed/common": "^6.62.0",
-    "@tsed/core": "^6.62.0",
-    "@tsed/exceptions": "^6.62.0",
-    "@tsed/logger": "^5.17.0",
-    "@tsed/schema": "^6.62.0",
-    "redis": "^3.1.2",
-    "uuid": "^8.3.2"
+    "@tsed/common": "^7.10.0",
+    "@tsed/core": "^7.10.0",
+    "@tsed/exceptions": "^7.10.0",
+    "@tsed/logger": "^6.3.3",
+    "@tsed/schema": "^7.10.0",
+    "redis": "^4.5.1",
+    "uuid": "^9.0.0"
   }
 }
