@@ -11,42 +11,42 @@
   "devDependencies": {
     "@ftl/types": "1.0.0",
     "@types/compression": "^1.7.2",
-    "@types/express": "^4.17.13",
+    "@types/express": "^4.17.15",
     "@types/jest": "^26.0.24",
     "@types/multer": "^1.4.7",
-    "@types/node": "^16.4.9",
-    "@typescript-eslint/eslint-plugin": "^4.31.1",
-    "@typescript-eslint/parser": "^4.31.1",
-    "eslint": "^7.32.0",
+    "@types/node": "^18.11.18",
+    "@typescript-eslint/eslint-plugin": "^5.48.0",
+    "@typescript-eslint/parser": "^5.48.0",
+    "eslint": "^8.31.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-plugin-import": "^2.24.2",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.25.2",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "ts-node-dev": "^1.1.8",
-    "typescript": "^4.3.5"
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^4.9.4"
   },
   "dependencies": {
     "@ftl/api": "1.0.0",
     "@ftl/common": "1.0.0",
     "@ftl/types": "1.0.0",
-    "@tsed/ajv": "^6.62.0",
-    "@tsed/common": "^6.62.0",
-    "@tsed/core": "^6.62.0",
-    "@tsed/di": "^6.62.0",
-    "@tsed/exceptions": "^6.62.0",
-    "@tsed/json-mapper": "^6.62.0",
-    "@tsed/logger": "^5.17.0",
-    "@tsed/mongoose": "^6.62.0",
-    "@tsed/platform-express": "^6.62.0",
-    "@tsed/schema": "^6.62.0",
-    "ajv": "^8.6.3",
+    "@tsed/ajv": "^7.10.0",
+    "@tsed/common": "^7.10.0",
+    "@tsed/core": "^7.10.0",
+    "@tsed/di": "^7.10.0",
+    "@tsed/exceptions": "^7.10.0",
+    "@tsed/json-mapper": "^7.10.0",
+    "@tsed/logger": "^6.3.3",
+    "@tsed/mongoose": "^7.10.0",
+    "@tsed/platform-express": "^7.10.0",
+    "@tsed/schema": "^7.10.0",
+    "ajv": "^8.12.0",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.5",
-    "express": "4",
+    "express": "^4.18.2",
     "express-session": "^1.17.2",
-    "msgpack5": "^5.3.2",
-    "uuid": "^8.3.2"
+    "mongoose": "^6.8.3",
+    "uuid": "^9.0.0"
   }
 }
