@@ -13,25 +13,25 @@
     "@types/express": "^4.17.15",
     "@types/express-ws": "^3.0.1",
     "@types/jest": "^26.0.24",
-    "@types/node": "^16.4.9",
+    "@types/node": "^18.11.18",
     "redis-mock": "^0.56.3",
     "ts-jest": "^29.0.3",
-    "ts-node-dev": "^1.1.8",
-    "typescript": "^4.3.5"
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^4.9.4"
   },
   "dependencies": {
     "@beyond/protocol": "^0.1.5",
     "@ftl/api": "1.0.0",
     "@ftl/common": "1.0.0",
     "@ftl/types": "1.0.0",
-    "@tsed/logger": "^5.17.0",
+    "@tsed/logger": "^6.3.3",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.5",
-    "express": "4",
+    "express": "^4.18.2",
     "express-ws": "^5.0.2",
-    "msgpack5": "^5.3.2",
+    "msgpack5": "^6.0.2",
     "uuid": "^9.0.0",
-    "ws": "^8.11.0"
+    "ws": "^8.12.0"
   }
 }
