@@ -1,5 +1,5 @@
 import React from 'react';
-import {Switch, Route} from 'react-router';
+import {Routes, Route} from 'react-router-dom';
 import {SKRView} from '../SKRView';
 import {DeveloperView} from '../Developer';
 import {GeneralView} from './GeneralView';
@@ -8,10 +8,10 @@ import {IStream} from '../../api/streams';
 export function Router({ data }: { data: IStream }) {
     const path = process.env.ASSET_PATH;
     return (
-        <Switch>
-            <Route exact path={`${path}view`} render={props => <GeneralView {...props} data={data} />} />
-            <Route path={`${path}view/skr`} component={SKRView} />
-            <Route path={`${path}view/developer`} component={DeveloperView} />
-        </Switch>
+        <Routes>
+            <Route path={`${path}*`} element={<GeneralView data={data}/>} />
+            <Route path={`${path}view/skr`} element={<SKRView />} />
+            <Route path={`${path}view/developer`} element={<DeveloperView />} />
+        </Routes>
     )
 }
