@@ -1,175 +1,134 @@
-import redis from 'redis';
+import * as redis from 'redis';
 import { v4 as uuidv4 } from 'uuid';
 import { $log } from '@tsed/logger';
 import { BaseEventBody, Event } from '@ftl/api';
 import * as fs from 'fs';
 import ALS from './als';
 
-let redisClient: redis.RedisClient;
-let redisSub: redis.RedisClient;
-let redisBlockClient: redis.RedisClient;
+let redisClient: redis.RedisClientType;
+let redisSub: redis.RedisClientType;
+let redisBlockClient: redis.RedisClientType;
 
-const handlers = new Map<string, Function[]>();
+let redisPromise: Promise<void>;
+
+// const handlers = new Map<string, Function[]>();
+const subscriptions = new Set<string>();
 
 const DAY = 1 * 24 * 60 * 60 * 1000;
 
-function initRedis(): void {
+async function initRedis() {
   if (!redisClient) {
-    redisClient = redis.createClient({
-      host: process.env.REDIS_HOST || 'redis',
-    });
-    redisSub = redisClient.duplicate({ return_buffers: true });
-    // redisPub = redisClient.duplicate();
-    redisBlockClient = redisClient.duplicate();
+    if (redisPromise) return redisPromise;
 
-    redisSub.on('message', (channel: Buffer, message: Buffer) => {
-      const strchan = channel.toString('utf8');
-
-      if (handlers.has(strchan)) {
-        const hs = handlers.get(strchan);
-        for (const h of hs) {
-          h(message);
-        }
-      }
-    });
-  }
-}
-
-export function redisPublish(channel: string, data: unknown): Promise<boolean> {
-  initRedis();
-  // Publish to redis
-  return new Promise((resolve) => {
-    redisClient.publish(channel, data as string, (err) => {
-      resolve(!err);
-    })
-  });
-}
-
-export function redisSubscribe(channel: string, cb: Function): Promise<boolean> {
-  initRedis();
-
-  const hasChannel = handlers.has(channel);
-  // Subscribe to redis
-  if (hasChannel) {
-    handlers.get(channel).push(cb);
-  } else {
-    handlers.set(channel, [cb]);
-  }
-  return new Promise((resolve) => {
-    if (!hasChannel) {
-      redisSub.subscribe(channel, (err) => {
-        if (err) {
-          $log.error('Subscribe error', err);
-        }
-        resolve(!err);
+    redisPromise = new Promise(async (resolve) => {
+      const tmpClient: redis.RedisClientType = redis.createClient({
+        url: `redis://${process.env.REDIS_HOST || 'redis'}`,
+        socket: {
+          host: process.env.REDIS_HOST || 'redis',
+        },
       });
-    }
-    return true;
-  });
+      tmpClient.on('error', (err) => {
+        console.log('Redis ERROR', err);
+      });
+      redisSub = tmpClient.duplicate();
+      redisSub.on('error', (err) => {
+        console.log('Redis Sub ERROR', err);
+      });
+      // redisPub = redisClient.duplicate();
+      redisBlockClient = tmpClient.duplicate();
+      redisBlockClient.on('error', (err) => {
+        console.log('Redis Block ERROR', err);
+      });
+  
+      /*redisSub.on('message', (channel: Buffer, message: Buffer) => {
+        $log.info('On message', channel);
+        const strchan = channel.toString('utf8');
+  
+        if (handlers.has(strchan)) {
+          const hs = handlers.get(strchan);
+          for (const h of hs) {
+            h(message);
+          }
+        }
+      });*/
+  
+      await tmpClient.connect();
+      await redisSub.connect();
+      await redisBlockClient.connect();
+      redisClient = tmpClient;
+      resolve();
+    });
+
+    return redisPromise;
+  }
 }
 
-export function redisUnsubscribe(channel: string, cb: Function) {
-  if (!handlers.has(channel)) {
+export async function redisPublish(channel: string, data: unknown): Promise<number> {
+  await initRedis();
+  // Publish to redis
+  return redisClient.publish(channel, data as string);
+}
+
+export async function redisSubscribe(channel: string, cb: (msg: Buffer) => void): Promise<void> {
+  await initRedis();
+
+  const hasChannel = subscriptions.has(channel);
+
+  if (!hasChannel) {
+    subscriptions.add(channel);
+    return redisSub.subscribe(channel, cb, true);
+  }
+}
+
+export async function redisUnsubscribe(channel: string, cb: (msg: Buffer) => void) {
+  if (!subscriptions.has(channel)) {
     return;
   }
 
-  const hs = handlers.get(channel);
-  const filtered = hs.filter((f) => f !== cb);
-
-  if (filtered.length === 0) {
-    handlers.delete(channel);
-  } else {
-    handlers.set(channel, filtered);
-  }
-
-  if (filtered.length === 0) {
-    redisSub.unsubscribe(channel, (err) => {
-      if (err) {
-        $log.error('Subscribe error', err);
-      } else {
-        $log.info('Remove redis subscription', channel);
-      }
-    });
-  }
+  subscriptions.delete(channel);
+  return redisSub.unsubscribe(channel, cb, true);
 }
 
-export function redisSet<T>(key: string, item: T, time?: number): Promise<boolean> {
-  initRedis();
+export async function redisSet<T>(key: string, item: T, time?: number): Promise<string> {
+  await initRedis();
   if (time) {
-    return new Promise((resolve) => {
-      redisClient.multi().set(key, JSON.stringify(item)).expire(key, time).exec((err) => {
-        resolve(!err);
-      });
-    });
+    await redisClient.multi().set(key, JSON.stringify(item)).expire(key, time).exec();
+    return '';
   }
 
-  return new Promise((resolve) => {
-    redisClient.set(key, JSON.stringify(item), (err) => {
-      resolve(!err);
-    });
-  });
+  return redisClient.set(key, JSON.stringify(item));
 }
 
-export function redisGet<T>(key: string): Promise<T> {
-  initRedis();
-  return new Promise((resolve) => {
-    redisClient.get(key, (err, reply) => {
-      if (err) {
-        resolve(null);
-      } else {
-        resolve(JSON.parse(reply))
-      }
-    });
-  });
+export async function redisGet<T>(key: string): Promise<T> {
+  await initRedis();
+  const result = await redisClient.get(key);
+  return JSON.parse(result)
 }
 
-export function redisMGet<T>(keys: string[]): Promise<T[]> {
-  initRedis();
-  return new Promise((resolve) => {
-    redisClient.mget(...keys, (err, reply) => {
-      if (err) {
-        resolve(null);
-      } else {
-        resolve(JSON.parse(reply))
-      }
-    });
-  });
+export async function redisMGet<T>(keys: string[]): Promise<T[]> {
+  await initRedis();
+  const result = await redisClient.mGet(keys)
+  return result.map((v) => JSON.parse(v));
 }
 
-export function redisDelete(key: string): Promise<boolean> {
-  initRedis();
-  return new Promise((resolve) => {
-    redisClient.del(key, (err) => {
-      resolve(!err);
-    });
-  });
+export async function redisDelete(key: string): Promise<number> {
+  await initRedis();
+  return redisClient.del(key);
 }
 
-export function redisAddItem(key: string, item: string, time: number): Promise<boolean> {
-  initRedis();
-  return new Promise((resolve) => {
-    redisClient.zadd(key, time, item, (err) => {
-      resolve(!err);
-    });
-  });
+export async function redisAddItem(key: string, item: string, time: number): Promise<number> {
+  await initRedis();
+  return redisClient.zAdd(key, { score: time, value: item });
 }
 
-export function redisRemoveItem(key: string, item: string): Promise<boolean> {
-  initRedis();
-  return new Promise((resolve) => {
-    redisClient.zrem(key, item, (err) => {
-      resolve(!err);
-    });
-  });
+export async function redisRemoveItem(key: string, item: string): Promise<number> {
+  await initRedis();
+  return redisClient.zRem(key, item);
 }
 
-export function redisTopItems(key: string): Promise<string[]> {
-  initRedis();
-  return new Promise((resolve) => {
-    redisClient.zrangebyscore(key, Date.now() - 1 * DAY, Date.now(), (err, reply) => {
-      resolve(err ? [] : reply);
-    });
-  });
+export async function redisTopItems(key: string): Promise<string[]> {
+  await initRedis();
+  return redisClient.zRangeByScore(key, Date.now() - 1 * DAY, Date.now());
 }
 
 // eslint-disable-next-line no-unused-vars
@@ -186,43 +145,34 @@ export function redisReply(cb: ReplyCallback): string {
 const originKey = `reply:${uuidv4()}`;
 
 redisSubscribe(originKey, (data) => {
-  const res = JSON.parse(data);
+  const res = JSON.parse(data.toString('utf8'));
   if (res?.request && replyMap.has(res.request)) {
     replyMap.get(res.request)(res.data);
     replyMap.delete(res.request);
   }
 });
 
-export function redisSendCommand(service: string, request: string, cmd: string, data: unknown) {
-  initRedis();
+export async function redisSendCommand(service: string, request: string, cmd: string, data: unknown) {
+  await initRedis();
   const key = `command:${service}`;
-  redisClient.rpush(key, JSON.stringify({
+  return redisClient.rPush(key, JSON.stringify({
     cmd, data, request, origin: originKey,
-  }), (err) => {
-    if (err) {
-      $log.error('Redis command send error', err);
-    }
-  });
+  }));
 }
 
 // eslint-disable-next-line no-unused-vars
 type CommandCallback = (cmd: string, data: unknown) => unknown;
 
-export function redisListen(service: string, cb: CommandCallback) {
-  initRedis();
+export async function redisListen(service: string, cb: CommandCallback) {
+  await initRedis();
   const key = `command:${service}`;
-  redisBlockClient.blpop(key, 0, async (err, rep) => {
-    if (err) {
-      $log.error('Command error', err);
-    } else {
-      const data = JSON.parse(rep[1]);
-      const result = await cb(data.cmd, data.data);
-      if (data.request && data.origin) {
-        redisPublish(data.origin, JSON.stringify({ data: result, request: data.request }));
-      }
-    }
-    redisListen(service, cb);
-  });
+  const rep = await redisBlockClient.blPop(key, 0);
+  const data = JSON.parse(rep.element);
+  const result = await cb(data.cmd, data.data);
+  if (data.request && data.origin) {
+    redisPublish(data.origin, JSON.stringify({ data: result, request: data.request }));
+  }
+  redisListen(service, cb);
 }
 
 function keysFromObject(obj: any): string[] {
@@ -255,12 +205,50 @@ function keysFromObject(obj: any): string[] {
   return result;
 }
 
+function itemsFromObject(obj: any): Record<string, string> {
+  const result = {};
+  // eslint-disable-next-line guard-for-in
+  for (const key in obj) {
+    const value = obj[key];
+
+    const type = typeof value;
+    let valueStr: string;
+    switch (type) {
+      case 'string':
+        valueStr = value;
+        break;
+      case 'number':
+        valueStr = `${value}`;
+        break;
+      case 'object':
+        valueStr = JSON.stringify(value);
+        break;
+      default:
+        break;
+    }
+
+    if (valueStr) {
+      result[key] = valueStr;
+    }
+  }
+  return result;
+}
+
 function objectFromKeys(keys: string[]): any {
   const obj = {};
   for (let i = 0; i < keys.length; i += 2) {
     const k = keys[i + 1];
     // eslint-disable-next-line no-restricted-globals
     obj[keys[i]] = (isNaN(k as any)) ? k : parseFloat(k);
+  }
+  return obj;
+}
+
+function objectTypeCorrection(obj: any): any {
+  for (const k of Object.keys(obj)) {
+    const v = obj[k];
+    // eslint-disable-next-line no-restricted-globals
+    obj[k] = (isNaN(v as any)) ? v : parseFloat(v);
   }
   return obj;
 }
@@ -276,35 +264,22 @@ function objectFromValues(keys: string[], values: any[]): any {
   return obj;
 }
 
-export function redisHSet(key: string, data: any, ttl: number): Promise<boolean> {
-  initRedis();
+export async function redisHSet(key: string, data: any, ttl: number) {
+  await initRedis();
   const items = keysFromObject(data);
-  return new Promise((resolve) => {
-    redisClient.multi().hset(
-      key, ...items,
-    ).expire(key, ttl).exec((err) => resolve(!err))
-  });
+  return redisClient.multi().hSet(
+      key, items,
+    ).expire(key, ttl).exec();
 }
 
-export function redisHGetM(key: string, items: string[]): Promise<any> {
-  initRedis();
-  return new Promise((resolve) => {
-    redisClient.hmget(
-      key, ...items,
-      (err, rep) => {
-        if (err) {
-          $log.error('Redis hmget error', err);
-          resolve(null);
-        } else {
-          resolve(objectFromValues(items, rep));
-        }
-      },
-    )
-  });
+export async function redisHGetM(key: string, items: string[]): Promise<any> {
+  await initRedis();
+  const result = await redisClient.hmGet(key, items);
+  return objectFromValues(items, result);
 }
 
-export function redisSendEvent<T extends Event>(event: T) {
-  initRedis();
+export async function redisSendEvent<T extends Event>(event: T) {
+  await initRedis();
   if (!event.body.operationId) {
     const opId = ALS.getStore() || new Map<string, string>();
     if (!opId.has('operationId')) {
@@ -319,19 +294,15 @@ export function redisSendEvent<T extends Event>(event: T) {
     // eslint-disable-next-line no-param-reassign
     event.body.clientId = opId.get('clientId');
   }
-  const items = keysFromObject(event.body);
-  if (!redisClient.xadd) return;
-  redisClient.xadd.apply(
-    redisClient,
-    [
-      event.event, 'MAXLEN', '~', '1000', '*',
-      ...items, (err) => {
-        if (err) {
-          $log.error('Redis event error', err);
-        }
-      },
-    ],
-  );
+  const items = itemsFromObject(event.body);
+  if (!redisClient.xAdd) return;
+  return redisClient.xAdd(event.event, '*', items, {
+    TRIM: {
+      strategy: 'MAXLEN',
+      strategyModifier: '~',
+      threshold: 1000,
+      limit: 1000,
+  }});
 }
 
 const streamKeys = new Set<string>();
@@ -339,28 +310,23 @@ let streamListen = false;
 // eslint-disable-next-line no-unused-vars
 type StreamCallback<T extends BaseEventBody> = (data: T, id: string) => Promise<void>;
 const streamCallbacks = new Map<string, StreamCallback<BaseEventBody>>();
-let streamID = '0';
 let consumerGroup = null;
 let consumerId = null;
+const readyStreams = new Set<string>();
+const streamIds = new Map<string, string>();
 
 export function redisSetGroup(group: string) {
   consumerGroup = group;
 }
 
-export async function redisCreateGroup(stream: string): Promise<number> {
-  return new Promise((resolve) => {
-    redisClient.xgroup('CREATE', stream, consumerGroup, '$', 'MKSTREAM', (err, rep) => {
-      if (err) {
-        $log.error('Redis group error', err);
-        resolve(-1);
-      } else {
-        resolve(rep);
-      }
-    });
-  });
+export async function redisCreateGroup(stream: string) {
+  await initRedis();
+  try {
+    await redisClient.xGroupCreate(stream, consumerGroup, '$', { MKSTREAM: true });
+  } catch(e) {}
 }
 
-export function redisSetStreamCallback<T extends Event>(key: T['event'], cb: StreamCallback<T['body']>): Promise<number> {
+export function redisSetStreamCallback<T extends Event>(key: T['event'], cb: StreamCallback<T['body']>) {
   streamCallbacks.set(key, cb);
   streamKeys.add(key);
   return redisCreateGroup(key);
@@ -387,13 +353,13 @@ export function redisConsumerId() {
   return name;
 }
 
-export function redisStreamListen(optName?: string, newKeys?: string[]) {
+export async function redisStreamListen(optName?: string, newKeys?: string[]) {
   let name = optName;
 
   if (!name) {
     name = redisConsumerId();
   }
-  initRedis();
+  await initRedis();
 
   if (Array.isArray(newKeys)) {
     newKeys.forEach((k) => streamKeys.add(k));
@@ -404,108 +370,97 @@ export function redisStreamListen(optName?: string, newKeys?: string[]) {
   }
   streamListen = true;
 
-  if ((!consumerGroup || !name) && streamID === '0') {
-    streamID = '$';
-  }
-
   const keys = Array.from(streamKeys);
-  const ids = keys.map(() => streamID);
 
   if (keys.length === 0) {
     return;
   }
 
   if (!consumerGroup || !name) {
-    redisBlockClient.xread.apply(
-      redisBlockClient,
-      ['BLOCK', '60000', 'STREAMS', ...keys, ...ids, async (err, rep) => {
-        if (err) {
-          $log.error('Command error', err);
-        } else if (Array.isArray(rep)) {
-          for (const stream of rep) {
-            const [key, dataSet] = stream;
-            for (const data of dataSet) {
-              const [id, columns] = data;
-              streamID = id;
-              const obj = objectFromKeys(columns) as BaseEventBody;
-              if (streamCallbacks.has(key)) {
-                if (obj.operationId) {
-                  const state = new Map<string, string>();
-                  state.set('operationId', obj.operationId);
-                  if (obj.userId) state.set('userId', obj.userId);
-                  if (obj.sessionId) state.set('sessionId', obj.sessionId);
-                  if (obj.clientId) state.set('clientId', obj.clientId);
-                  ALS.run(state, () => {
-                    streamCallbacks.get(key)(obj, id).catch((e) => {
-                      $log.error('Stream callback error', e);
-                    });
-                  });
-                } else {
-                  streamCallbacks.get(key)(obj, id).catch((e) => {
-                    $log.error('Stream callback error', e);
-                  });
-                }
-              }
+    const pairs = keys.map((v) => ({ key: v, id: (streamIds.has(v)) ? streamIds.get(v) : '$'}));
+    const rep = await redisBlockClient.xRead(pairs, { BLOCK: 60000 });
+
+    if (Array.isArray(rep)) {
+      for (const stream of rep) {
+        const streamName = stream.name;
+
+        for (const msg of stream.messages) {
+          streamIds.set(streamName, msg.id);
+          if (streamCallbacks.has(streamName)) {
+            const obj = objectTypeCorrection(msg.message) as unknown as BaseEventBody;
+
+            if (obj.operationId) {
+              const state = new Map<string, string>();
+              state.set('operationId', obj.operationId);
+              if (obj.userId) state.set('userId', obj.userId);
+              if (obj.sessionId) state.set('sessionId', obj.sessionId);
+              if (obj.clientId) state.set('clientId', obj.clientId);
+              // eslint-disable-next-line no-loop-func
+              ALS.run(state, () => {
+                // eslint-disable-next-line no-loop-func
+                streamCallbacks.get(streamName)(obj, msg.id).catch((e) => {
+                  $log.error('Stream callback error', e);
+                });
+              });
+            } else {
+              // eslint-disable-next-line no-loop-func
+              streamCallbacks.get(streamName)(obj, msg.id).catch((e) => {
+                $log.error('Stream callback error', e);
+              });
             }
           }
         }
-        streamListen = false;
-        redisStreamListen(name);
-      }],
-    );
+      }
+    }
+    streamListen = false;
+    redisStreamListen(name);
   } else {
-    redisBlockClient.xreadgroup.apply(
-      redisBlockClient,
-      ['GROUP', consumerGroup, name, 'BLOCK', '60000', 'STREAMS', ...keys, ...ids, async (err, rep) => {
-        if (err) {
-          $log.error('Command error', err);
-        } else if (Array.isArray(rep)) {
-          for (const stream of rep) {
-            const [key, dataSet] = stream;
-            for (const data of dataSet) {
-              const [id, columns] = data;
-              streamID = id;
-              const obj = objectFromKeys(columns);
-              if (streamCallbacks.has(key)) {
-                if (obj.operationId) {
-                  const state = new Map<string, string>();
-                  state.set('operationId', obj.operationId);
-                  if (obj.userId) state.set('userId', obj.userId);
-                  if (obj.sessionId) state.set('sessionId', obj.sessionId);
-                  if (obj.clientId) state.set('clientId', obj.clientId);
-                  // eslint-disable-next-line no-loop-func
-                  ALS.run(state, () => {
-                    // eslint-disable-next-line no-loop-func
-                    streamCallbacks.get(key)(obj, id).then(() => {
-                      redisClient.xack(key, consumerGroup, id, (ackerr) => {
-                        if (err) {
-                          $log.error('ACK Error', ackerr);
-                        }
-                      });
-                    }).catch((e) => {
-                      $log.error('Stream callback error', e);
-                    });
-                  });
-                } else {
-                  // eslint-disable-next-line no-loop-func
-                  streamCallbacks.get(key)(obj, id).then(() => {
-                    redisClient.xack(key, consumerGroup, id, (ackerr) => {
-                      if (err) {
-                        $log.error('ACK Error', ackerr);
-                      }
-                    });
-                  }).catch((e) => {
-                    $log.error('Stream callback error', e);
-                  });
-                }
-              }
+    const pairs = keys.map((v) => ({ key: v, id: (readyStreams.has(v)) ? '>' : '0'}));
+    const rep = await redisBlockClient.xReadGroup(consumerGroup, name, pairs, {
+      BLOCK: 20000,
+    });
+
+    if (Array.isArray(rep)) {
+      for (const stream of rep) {
+        const streamName = (typeof stream.name === 'string') ? stream.name : stream.name.toString('utf8');
+
+        if (stream.messages.length === 0) {
+          readyStreams.add(streamName);
+        }
+
+        for (const msg of stream.messages) {
+          if (streamCallbacks.has(streamName)) {
+            const msgId = (typeof msg.id === 'string') ? msg.id : msg.id.toString('utf8');
+            const obj = objectTypeCorrection(msg.message) as unknown as BaseEventBody;
+
+            if (obj.operationId) {
+              const state = new Map<string, string>();
+              state.set('operationId', obj.operationId);
+              if (obj.userId) state.set('userId', obj.userId);
+              if (obj.sessionId) state.set('sessionId', obj.sessionId);
+              if (obj.clientId) state.set('clientId', obj.clientId);
+              // eslint-disable-next-line no-loop-func
+              ALS.run(state, () => {
+                // eslint-disable-next-line no-loop-func
+                streamCallbacks.get(streamName)(obj, msgId).then(() => {
+                  return redisClient.xAck(streamName, consumerGroup, msg.id);
+                }).catch((e) => {
+                  $log.error('Stream callback error', e);
+                });
+              });
+            } else {
+              // eslint-disable-next-line no-loop-func
+              streamCallbacks.get(streamName)(obj, msgId).then(() => {
+                return redisClient.xAck(streamName, consumerGroup, msg.id);
+              }).catch((e) => {
+                $log.error('Stream callback error', e);
+              });
             }
           }
         }
-        streamListen = false;
-        streamID = '>';
-        redisStreamListen(name);
-      }],
-    );
+      }
+    }
+    streamListen = false;
+    redisStreamListen(name);
   }
 }
