@@ -1,12 +1,14 @@
 import {
-  Configuration, HttpServer, Inject, PlatformApplication,
+  Configuration, Inject, PlatformApplication,
 } from '@tsed/common';
 import { $log } from '@tsed/logger';
 import express from 'express';
 import compress from 'compression';
 import cookieParser from 'cookie-parser';
 import { redisStreamListen, redisSetGroup, installMonitor } from '@ftl/common';
+import Http from 'http';
 import './logger';
+import Nodes from './controllers/nodes';
 
 $log.appenders.set('redis', {
   type: 'redis',
@@ -21,7 +23,7 @@ const rootDir = __dirname;
   port: 8080,
   debug: false,
   mount: {
-    '/v1': `${rootDir}/controllers/**/*.ts`,
+    '/v1': [Nodes],
   },
   logger: {
     logRequest: false,
@@ -39,8 +41,8 @@ export default class Server {
   @Inject()
   app: PlatformApplication;
 
-  @Inject(HttpServer)
-  httpServer: HttpServer;
+  @Inject(Http.Server)
+  httpServer: Http.Server;
 
   @Configuration()
   settings: Configuration;
